+++ RELEASE_NOTES.md
@@ -1,8 +1,12 @@
 Release Notes
 =============
+## 1.2.0-beta1
+* Bastion Hosts: Create bastion hosts for accessing resources on a virtual network.
+
 ## 1.1.0
 * Cognitive Services: Retrieve ARM expression to the Key of the Cognitive Services instance.
 * Web App / Functions: Allow CORS enable credentials (https://github.com/CompositionalIT/farmer/issues/265)
+* Gateway: Add VPN Client configuration
 * Container Instance: Change modelling from an anonymous type to a discriminated union (interop) (https://github.com/CompositionalIT/farmer/issues/372)
 * Fix a bug whereby optional Location, Tags and DependsOn were set to empty lists instead of null when not required.
 * Storage: Create connection string expressions.